feat:show all products

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\ProductController;
+use App\Http\Controllers\UserController;
 use App\Http\Middleware\IsAdmin;
 use Illuminate\Support\Facades\Route;
 
@@ -23,21 +24,18 @@ Route::middleware([
 Route::get('redirect',[HomeController::class,'redirect']);
 
 
-Route::controller(ProductController::class)->group(function () {
-    Route::middleware(IsAdmin::class)->group(function(){
-
-    Route::post('/products', 'store')->name('store');
-    Route::get('/products/create', 'create')->name('create');
-    Route::get('/products', 'allProducts')->name('allProducts');
-    Route::get('/products/show/{id}', 'show')->name('show');
-    Route::get('/products/edit/{id}', 'edit')->name('edit');
-    Route::put('/products/{id}', 'update')->name('update');
-    Route::delete('/products/{id}', 'delete')->name('delete');
-
-
+Route::prefix('admin')->middleware(IsAdmin::class)->group(function () {
+    Route::controller(ProductController::class)->group(function () {
+        Route::post('/products', 'store')->name('admin.products.store');
+        Route::get('/products/create', 'create')->name('admin.products.create');
+        Route::get('/products', 'allProducts')->name('admin.products.all');
+        Route::get('/products/show/{id}', 'show')->name('admin.products.show');
+        Route::get('/products/edit/{id}', 'edit')->name('admin.products.edit');
+        Route::put('/products/{id}', 'update')->name('admin.products.update');
+        Route::delete('/products/{id}', 'delete')->name('admin.products.delete');
+    });
 });
 
-});
 
 Route::get('change/{lang}',function($lang){
     if($lang=="ar"){
@@ -48,3 +46,19 @@ Route::get('change/{lang}',function($lang){
     }
     return redirect()->back();
 })->middleware('ChangeLang');
+
+Route::controller(UserController::class)->group(function () {
+    // Route::middleware(IsAdmin::class)->group(function(){
+
+    Route::post('/products', 'store')->name('store');
+    Route::get('/products/create', 'create')->name('create');
+    Route::get('/products', 'allProducts')->name('allProducts');
+    Route::get('/products/show/{id}', 'show')->name('show');
+    Route::get('/products/edit/{id}', 'edit')->name('edit');
+    Route::put('/products/{id}', 'update')->name('update');
+    Route::delete('/products/{id}', 'delete')->name('delete');
+
+
+// });
+
+});
